refactor(icp): move identity loading into icp-cli
Loading an identity means reading PEM files, talking to the OS keyring or
an HSM, and — when the key is encrypted — asking the user for a password.
None of that belongs in a library whose job is to build and deploy: it
should be handed an already-constructed identity or agent.

`crates/icp/src/identity/` therefore moves to `crates/icp-cli/src/identity/`,
and with it the two pieces of `Context` that only existed to serve it:

- `password_func` is deleted from the library. 26 of its 37 references
  were inside the identity module; the rest were wiring. No `Prompt` port
  replaces it — with identity in the frontend there is nothing to abstract.
- `identity: Arc<dyn identity::Load>` is deleted, together with the
  selection-resolving helpers built on it (`get_identity`, `get_agent`,
  `get_agent_for_{env,network,url}`). Those take an `IdentitySelection`,
  a CLI concept, and must interleave identity loading with root-key
  resolution, so they now live on the CLI's `Context`. What stays in the
  library is `Context::create_agent`, which turns an identity plus a
  resolved network into an `Agent`.

The two genuine couplings outside the module:

- `telemetry_data.rs`: `IdentityStorageType` stays (telemetry is a
  library-side data bag); its `From<&IdentitySpec>` conversion travels
  with `IdentitySpec` into the CLI.
- `directories.rs`: `Access::identity()` becomes `identity_dir()`,
  returning a plain path. `IdentityPaths`/`IdentityDirectories` travel
  with identity; the generic directory-lock machinery (`icp::fs::lock`)
  stays, exactly as it does for settings and the package cache.

`icp::context::initialize` no longer builds an identity loader, so it can
no longer fail: it returns a `Context` and `ContextInitError` is gone
from the library. The `ICP_CLI_KEYRING_MOCK_DIR` test hook moves to the
CLI's `initialize` alongside the loader it configures.

Dependencies that leave `icp` entirely (74 -> 58): bip32, crypto-bigint,
elliptic-curve, hmac, ic-ed25519, ic-identity-hsm, k256, keyring, p256,
pem, pkcs8, rand, scrypt, sec1, tiny-bip39, zeroize. `keyring`'s only
non-identity user was the mock hook above, so it leaves too.

Filesystem call sites in `crates/icp/src`: 241 -> 137.

The `Context` tests split along the same line: identity and agent
resolution move to `crates/icp-cli/src/context/tests.rs`, the rest stay.
So that the CLI can build a mocked library context, `icp`'s port mocks
are now compiled under `#[cfg(any(test, feature = "mocks"))]` and
`icp-cli` enables that feature as a dev-dependency.

Three dead items surfaced when the module landed in a binary crate, where
`pub` no longer suppresses dead-code analysis, and are removed:
`load_identity_in_context`, `IdentityPaths::ensure_identity_{defaults,list}_path`,
`WriteIdentityError::CreateDirectoryError` and `MockIdentityLoader::with_default`.
