+++ crates/icp-cli/Cargo.toml
@@ -29,6 +29,7 @@ cargo-generate.workspace = true
 clap-markdown.workspace = true
 clap.workspace = true
 clap_complete.workspace = true
+crypto-bigint.workspace = true
 dialoguer.workspace = true
 dunce.workspace = true
 elliptic-curve.workspace = true
@@ -37,9 +38,11 @@ futures.workspace = true
 glob.workspace = true
 tar.workspace = true
 hex.workspace = true
+hmac.workspace = true
 httptest.workspace = true
 ic-agent.workspace = true
 ic-ed25519.workspace = true
+ic-identity-hsm.workspace = true
 ic-ledger-types.workspace = true
 ic-management-canister-types.workspace = true
 ic-utils.workspace = true
@@ -52,6 +55,7 @@ indicatif.workspace = true
 indoc.workspace = true
 itertools.workspace = true
 k256.workspace = true
+keyring.workspace = true
 lazy_static.workspace = true
 num-bigint.workspace = true
 num-integer.workspace = true
@@ -65,6 +69,7 @@ pkcs8.workspace = true
 rand.workspace = true
 regex.workspace = true
 reqwest.workspace = true
+scrypt.workspace = true
 sec1.workspace = true
 semver.workspace = true
 serde_json.workspace = true
@@ -73,6 +78,7 @@ serde.workspace = true
 sha2.workspace = true
 shellwords.workspace = true
 snafu.workspace = true
+strum.workspace = true
 sysinfo.workspace = true
 tiny-bip39.workspace = true
 time.workspace = true
@@ -83,6 +89,7 @@ url.workspace = true
 uuid.workspace = true
 wasmparser.workspace = true
 wslpath2.workspace = true
+zeroize.workspace = true
 
 [target.'cfg(unix)'.dependencies]
 cargo-generate = { workspace = true, features = ["vendored-openssl"] }
@@ -91,6 +98,7 @@ cargo-generate = { workspace = true, features = ["vendored-openssl"] }
 assert_cmd.workspace = true
 camino-tempfile.workspace = true
 cryptoki.workspace = true
+icp = { workspace = true, features = ["clap", "mocks"] }
 predicates.workspace = true
 rand.workspace = true
 send_ctrlc.workspace = true
