+++ crates/icp-cli/src/commands/canister/create.rs
@@ -1,6 +1,7 @@
 use crate::context::Context;
 use std::io::stdout;
 
+use crate::identity::IdentitySelection;
 use anyhow::anyhow;
 use bigdecimal::BigDecimal;
 use candid::{Nat, Principal};
@@ -8,7 +9,6 @@ use clap::{ArgGroup, Args, Parser};
 use ic_management_canister_types::CanisterSettings as MgmtCanisterSettings;
 use icp::canister::resolve_controllers;
 use icp::context::{EnvironmentSelection, NetworkSelection};
-use icp::identity::IdentitySelection;
 use icp::parsers::{CyclesAmount, DurationAmount, MemoryAmount, parse_token_amount};
 use icp::store_id::IdMapping;
 use icp::{Canister, context::CanisterSelection, prelude::*};
