+++ crates/icp-cli/src/commands/identity/delegation/request.rs
@@ -1,8 +1,9 @@
 use crate::context::Context;
+use crate::identity::key;
 use clap::{Args, ValueHint};
 use dialoguer::Password;
 use elliptic_curve::zeroize::Zeroizing;
-use icp::{fs::read_to_string, identity::key, prelude::*};
+use icp::{fs::read_to_string, prelude::*};
 use pem::Pem;
 use snafu::{ResultExt, Snafu};
 use tracing::warn;
@@ -52,8 +53,7 @@ pub(crate) async fn exec(ctx: &Context, args: &RequestArgs) -> Result<(), Reques
     };
 
     let der_public_key = ctx
-        .dirs
-        .identity()?
+        .identity_dirs()?
         .with_write(async |dirs| key::create_pending_delegation(dirs, &args.name, create_format))
         .await?
         .context(CreateSnafu)?;
