+++ crates/icp-cli/src/commands/identity/export.rs
@@ -1,11 +1,11 @@
 use crate::context::Context;
+use crate::identity::key::{ExportFormat, export_identity};
 use anyhow::Context as _;
 use clap::{Args, ValueHint};
 use clap_complete::ArgValueCandidates;
 use dialoguer::Password;
 use elliptic_curve::zeroize::Zeroizing;
 use icp::fs::read_to_string;
-use icp::identity::key::{ExportFormat, export_identity};
 use icp::prelude::*;
 
 /// Print the PEM file for the identity
@@ -29,7 +29,7 @@ pub(crate) struct ExportArgs {
 }
 
 pub(crate) async fn exec(ctx: &Context, args: &ExportArgs) -> Result<(), anyhow::Error> {
-    let dirs = ctx.dirs.identity()?;
+    let dirs = ctx.identity_dirs()?;
 
     // Read password if necessary
     let export_format = if args.encrypt {
