+++ crates/icp-cli/src/commands/identity/rename.rs
@@ -1,7 +1,7 @@
 use crate::context::Context;
+use crate::identity::key::rename_identity;
 use clap::Args;
 use clap_complete::ArgValueCandidates;
-use icp::identity::key::rename_identity;
 use tracing::info;
 
 /// Rename an identity
@@ -16,7 +16,7 @@ pub(crate) struct RenameArgs {
 }
 
 pub(crate) async fn exec(ctx: &Context, args: &RenameArgs) -> Result<(), anyhow::Error> {
-    let dirs = ctx.dirs.identity()?;
+    let dirs = ctx.identity_dirs()?;
 
     dirs.with_write(async |dirs| {
         rename_identity(dirs, &args.old_name, &args.new_name)?;
