+++ crates/icp-cli/src/commands/network/ping.rs
@@ -1,8 +1,8 @@
 use crate::context::Context;
+use crate::identity::IdentitySelection;
 use anyhow::bail;
 use clap::Args;
 use ic_agent::{Agent, agent::status::Status};
-use icp::identity::IdentitySelection;
 use std::time::Duration;
 use tokio::time::sleep;
 use tracing::info;
