+++ crates/icp-cli/src/main.rs
@@ -19,6 +19,7 @@ mod complete;
 mod context;
 mod dist;
 mod events;
+mod identity;
 mod logging;
 mod manifest;
 pub(crate) mod operations;
@@ -165,7 +166,7 @@ async fn run() -> Result<(), Error> {
         "Starting icp-cli"
     );
 
-    let password_func: icp::identity::PasswordFunc = match cli.identity_password_file {
+    let password_func: identity::PasswordFunc = match cli.identity_password_file {
         Some(path) => Arc::new(move || {
             icp::fs::read_to_string(&path)
                 .map(|s| s.trim().to_string())
