+++ crates/icp/src/canister/sync/mod.rs
@@ -76,12 +76,12 @@ impl Synchronize for Syncer {
     }
 }
 
-#[cfg(test)]
+#[cfg(any(test, feature = "mocks"))]
 /// Unimplemented mock implementation of `Synchronize`.
 /// All methods panic with `unimplemented!()` when called.
 pub struct UnimplementedMockSyncer;
 
-#[cfg(test)]
+#[cfg(any(test, feature = "mocks"))]
 #[async_trait]
 impl Synchronize for UnimplementedMockSyncer {
     async fn sync(
