+++ crates/icp/src/context/init.rs
@@ -1,22 +1,11 @@
-use std::{sync::Arc, time::Duration};
-
-use snafu::prelude::*;
+use std::sync::Arc;
 
 use crate::canister::build::Builder;
 use crate::canister::sync::Syncer;
 use crate::context::Context;
 use crate::store_artifact::ArtifactStore;
 
-use crate::{
-    ProjectLoad, agent, identity, identity::PasswordFunc, manifest::ProjectRootLocate, network,
-    store_id,
-};
-
-#[derive(Debug, Snafu)]
-pub enum ContextInitError {
-    #[snafu(display("failed to lock identity directory"))]
-    IdentityDirectory { source: crate::fs::lock::LockError },
-}
+use crate::{ProjectLoad, agent, manifest::ProjectRootLocate, network, store_id};
 
 /// Assembles the library context from the ports the host provides: where its
 /// data lives, how to find a project, and how to load one.
@@ -24,9 +13,7 @@ pub fn initialize(
     dirs: Arc<dyn crate::directories::Access>,
     project_root_locate: Arc<dyn ProjectRootLocate>,
     project: Arc<dyn ProjectLoad>,
-    password_func: PasswordFunc,
-    pem_session_duration: Option<Duration>,
-) -> Result<Context, ContextInitError> {
+) -> Context {
     // Canister ID Store
     let ids = Arc::new(store_id::AccessImpl::new(project_root_locate.clone()));
 
@@ -42,22 +29,6 @@ pub fn initialize(
     // Telemetry data bag (written by subsystems, read at session finish)
     let telemetry_data = Arc::new(crate::telemetry_data::TelemetryData::default());
 
-    // Identity loader
-    let idload = Arc::new(identity::Loader::new(
-        dirs.identity().context(IdentityDirectorySnafu)?,
-        password_func.clone(),
-        pem_session_duration,
-        telemetry_data.clone(),
-    ));
-
-    if let Ok(mockdir) = std::env::var("ICP_CLI_KEYRING_MOCK_DIR") {
-        keyring::set_default_credential_builder(Box::new(
-            crate::identity::keyring_mock::MockKeyring {
-                dir: crate::prelude::PathBuf::from(mockdir),
-            },
-        ));
-    }
-
     // Agent creator
     let agent_creator = Arc::new(agent::Creator);
 
@@ -68,18 +39,15 @@ pub fn initialize(
         agent: agent_creator.clone(),
     });
 
-    // Setup environment
-    Ok(Context {
+    Context {
         dirs,
         ids,
         artifacts,
         project,
-        identity: idload,
         network: netaccess,
         agent: agent_creator,
         builder,
         syncer,
         telemetry_data,
-        password_func,
-    })
+    }
 }
