+++ crates/icp/src/context/mod.rs
@@ -6,7 +6,6 @@ use crate::{
     agent::CreateAgentError,
     canister::{build::Build, sync::Synchronize},
     directories,
-    identity::IdentitySelection,
     manifest::network::RootKeySpec,
     network::{Configuration as NetworkConfiguration, access::NetworkAccess},
     prelude::*,
@@ -19,7 +18,7 @@ use snafu::{OptionExt, ResultExt, Snafu};
 
 mod init;
 
-pub use init::{ContextInitError, initialize};
+pub use init::initialize;
 
 pub const IC_ROOT_KEY: &[u8; 133] = b"\x30\x81\x82\x30\x1d\x06\x0d\x2b\x06\x01\x04\x01\x82\xdc\x7c\x05\x03\x01\x02\x01\x06\x0c\x2b\x06\x01\x04\x01\x82\xdc\x7c\x05\x03\x02\x01\x03\x61\x00\x81\x4c\x0e\x6e\xc7\x1f\xab\x58\x3b\x08\xbd\x81\x37\x3c\x25\x5c\x3c\x37\x1b\x2e\x84\x86\x3c\x98\xa4\xf1\xe0\x8b\x74\x23\x5d\x14\xfb\x5d\x9c\x0c\xd5\x46\xd9\x68\x5f\x91\x3a\x0c\x0b\x2c\xc5\x34\x15\x83\xbf\x4b\x43\x92\xe4\x67\xdb\x96\xd6\x5b\x9b\xb4\xcb\x71\x71\x12\xf8\x47\x2e\x0d\x5a\x4d\x14\x50\x5f\xfd\x74\x84\xb0\x12\x91\x09\x1c\x5f\x87\xb9\x88\x83\x46\x3f\x98\x09\x1a\x0b\xaa\xae";
 
@@ -84,14 +83,11 @@ pub struct Context {
     /// Project loader
     pub project: Arc<dyn crate::ProjectLoad>,
 
-    /// Identity loader
-    identity: Arc<dyn crate::identity::Load>,
-
     /// NetworkAccess loader
     pub network: Arc<dyn crate::network::Access>,
 
     /// Agent creator
-    agent: Arc<dyn crate::agent::Create>,
+    pub agent: Arc<dyn crate::agent::Create>,
 
     /// Canister builder
     pub builder: Arc<dyn Build>,
@@ -101,27 +97,9 @@ pub struct Context {
 
     /// Telemetry data collected during command execution
     pub telemetry_data: Arc<crate::telemetry_data::TelemetryData>,
-
-    /// Password reader for identity decryption; shared with the identity loader.
-    pub password_func: Arc<dyn Fn() -> Result<String, String> + Send + Sync>,
 }
 
 impl Context {
-    /// Gets an identity based on the provided identity selection.
-    // TODO: refactor the whole codebase to use this method instead of directly accessing `ctx.identity.load()`
-    pub async fn get_identity(
-        &self,
-        identity: &IdentitySelection,
-        network_root_key: Option<Vec<u8>>,
-    ) -> Result<Arc<dyn Identity>, GetIdentityError> {
-        self.identity
-            .load(identity.clone(), network_root_key)
-            .await
-            .context(IdentityLoadSnafu {
-                identity: identity.clone(),
-            })
-    }
-
     /// Gets an environment by name from the currently loaded project.
     ///
     /// # Errors
@@ -372,38 +350,11 @@ impl Context {
         Ok(())
     }
 
-    /// Creates an agent for a given identity and environment.
-    pub async fn get_agent_for_env(
-        &self,
-        identity: &IdentitySelection,
-        environment: &EnvironmentSelection,
-    ) -> Result<Agent, GetAgentForEnvError> {
-        let env = self.get_environment(environment).await?;
-        let access = self.network.access(&env.network).await?;
-        let id = self
-            .get_identity(identity, Some(access.root_key.clone()))
-            .await?;
-        Ok(self.create_agent(id, access).await?)
-    }
-
-    /// Creates an agent for a given identity and network.
-    pub async fn get_agent_for_network(
-        &self,
-        identity: &IdentitySelection,
-        network_selection: &NetworkSelection,
-    ) -> Result<Agent, GetAgentForNetworkError> {
-        let network = self.get_network(network_selection).await?;
-        let access = self.network.access(&network).await?;
-        let id = self
-            .get_identity(identity, Some(access.root_key.clone()))
-            .await?;
-        Ok(self.create_agent(id, access).await?)
-    }
-
-    /// Private helper to create an agent given identity and network access.
+    /// Creates an agent for an identity, bound to an already resolved network.
     ///
-    /// Used by [`Self::get_agent_for_env`] and [`Self::get_agent_for_network`].
-    async fn create_agent(
+    /// The caller loads the identity: which identity to use, and how to unlock
+    /// it, is the frontend's business.
+    pub async fn create_agent(
         &self,
         id: Arc<dyn Identity>,
         network_access: NetworkAccess,
@@ -414,58 +365,6 @@ impl Context {
             .await?;
         agent.set_root_key(network_access.root_key);
         Ok(agent)
-    }
-
-    /// Creates an agent for a given identity and url.
-    pub async fn get_agent_for_url(
-        &self,
-        identity: &IdentitySelection,
-        url: &Url,
-    ) -> Result<Agent, GetAgentForUrlError> {
-        let id = self.get_identity(identity, None).await?;
-        let agent = self.agent.create(id, url.as_str()).await?;
-        Ok(agent)
-    }
-
-    pub async fn get_agent(
-        &self,
-        identity: &IdentitySelection,
-        network: &NetworkSelection,
-        environment: &EnvironmentSelection,
-    ) -> Result<Agent, GetAgentError> {
-        match (environment, network) {
-            // Error: Both environment and network specified
-            (EnvironmentSelection::Named(_), NetworkSelection::Named(_))
-            | (EnvironmentSelection::Named(_), NetworkSelection::Url(_, _)) => {
-                Err(GetAgentError::EnvironmentAndNetworkSpecified)
-            }
-
-            // Default environment + default network
-            (EnvironmentSelection::Default, NetworkSelection::Default) => {
-                // Try to get agent from the default environment if project exists
-                match self.get_agent_for_env(identity, environment).await {
-                    Ok(agent) => Ok(agent),
-                    Err(GetAgentForEnvError::GetEnvironment {
-                        source:
-                            GetEnvironmentError::ProjectLoad {
-                                source: crate::ProjectLoadError::Locate { .. },
-                            },
-                    }) => Err(GetAgentError::NoProjectOrNetwork),
-                    Err(e) => Err(e.into()),
-                }
-            }
-
-            // Environment specified
-            (EnvironmentSelection::Named(_), NetworkSelection::Default) => {
-                Ok(self.get_agent_for_env(identity, environment).await?)
-            }
-
-            // Network specified
-            (EnvironmentSelection::Default, NetworkSelection::Named(_))
-            | (EnvironmentSelection::Default, NetworkSelection::Url(_, _)) => {
-                Ok(self.get_agent_for_network(identity, network).await?)
-            }
-        }
     }
 
     pub async fn get_canister_id(
@@ -595,7 +494,7 @@ impl Context {
         }
     }
 
-    #[cfg(test)]
+    #[cfg(any(test, feature = "mocks"))]
     /// Creates a test context with all mocks
     pub fn mocked() -> Context {
         Context {
@@ -603,24 +502,13 @@ impl Context {
             ids: Arc::new(crate::store_id::mock::MockInMemoryIdStore::new()),
             artifacts: Arc::new(crate::store_artifact::MockInMemoryArtifactStore::new()),
             project: Arc::new(crate::MockProjectLoader::minimal()),
-            identity: Arc::new(crate::identity::MockIdentityLoader::anonymous()),
             network: Arc::new(crate::network::MockNetworkAccessor::new()),
             agent: Arc::new(crate::agent::Creator),
             builder: Arc::new(crate::canister::build::UnimplementedMockBuilder),
             syncer: Arc::new(crate::canister::sync::UnimplementedMockSyncer),
             telemetry_data: Arc::new(crate::telemetry_data::TelemetryData::default()),
-            password_func: Arc::new(|| Err("no password available in mock context".to_string())),
         }
     }
-}
-
-#[derive(Debug, Snafu)]
-pub enum GetIdentityError {
-    #[snafu(display("failed to load identity"))]
-    IdentityLoad {
-        source: crate::identity::LoadError,
-        identity: IdentitySelection,
-    },
 }
 
 #[derive(Debug, Snafu)]
@@ -733,74 +621,6 @@ pub enum RemoveCanisterIdForEnvError {
         canister_name: String,
         environment_name: String,
     },
-}
-
-#[derive(Debug, Snafu)]
-pub enum GetAgentForEnvError {
-    #[snafu(transparent)]
-    GetIdentity { source: GetIdentityError },
-
-    #[snafu(transparent)]
-    GetEnvironment { source: GetEnvironmentError },
-
-    #[snafu(transparent)]
-    NetworkAccess { source: crate::network::AccessError },
-
-    #[snafu(transparent)]
-    AgentCreate {
-        source: crate::agent::CreateAgentError,
-    },
-}
-
-#[derive(Debug, Snafu)]
-pub enum GetAgentForNetworkError {
-    #[snafu(transparent)]
-    GetIdentity { source: GetIdentityError },
-
-    #[snafu(transparent)]
-    GetNetwork { source: GetNetworkError },
-
-    #[snafu(transparent)]
-    NetworkAccess { source: crate::network::AccessError },
-
-    #[snafu(transparent)]
-    AgentCreate {
-        source: crate::agent::CreateAgentError,
-    },
-}
-
-#[derive(Debug, Snafu)]
-pub enum GetAgentForUrlError {
-    #[snafu(transparent)]
-    GetIdentity { source: GetIdentityError },
-
-    #[snafu(transparent)]
-    AgentCreate {
-        source: crate::agent::CreateAgentError,
-    },
-}
-
-#[derive(Debug, Snafu)]
-pub enum GetAgentError {
-    #[snafu(transparent)]
-    ProjectExists { source: crate::ProjectLoadError },
-
-    #[snafu(display("You can't specify both an environment and a network"))]
-    EnvironmentAndNetworkSpecified,
-
-    #[snafu(display(
-        "No project found and no network specified. Either run this command inside a project or specify a network with --network"
-    ))]
-    NoProjectOrNetwork,
-
-    #[snafu(transparent)]
-    GetAgentForEnv { source: GetAgentForEnvError },
-
-    #[snafu(transparent)]
-    GetAgentForNetwork { source: GetAgentForNetworkError },
-
-    #[snafu(transparent)]
-    GetAgentForUrl { source: GetAgentForUrlError },
 }
 
 #[derive(Debug, Snafu)]
